chore(front): Move mapHttpError to apps/frontend
