+++ apps/frontend/src/app/util/rxjs/map-http-error.spec.ts
