+++ apps/frontend/src/app/util/rxjs/map-http-error.ts
