+++ libs/util-fn/src/index.ts
@@ -11,6 +11,5 @@ export * from './array-from';
 export * from './kebab-case';
 export * from './deep-equal';
 export * from './extract-prefix-from-map-name';
-export * from './rxjs/map-http-error';
 export * from './leaderboard-key';
 export * from './omit';
